create bugs api,added new migrations

+++ app/database/migrations/2014_09_30_092731_create_project_user_table.php
@@ -12,15 +12,20 @@ class CreateProjectUserTable extends Migration {
 	 */
 	public function up()
 	{
-		Schema::create('project_user', function(Blueprint $table)
-		{
-			$table->increments('id');
-			$table->integer('project_id')->unsigned()->index();
-			$table->foreign('project_id')->references('id')->on('projects')->onDelete('cascade');
-			$table->integer('user_id')->unsigned()->index();
-			$table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-			$table->timestamps();
-		});
+        if (!Schema::hasTable('project_user'))
+        {
+            Schema::create('project_user', function(Blueprint $table)
+            {
+                $table->increments('id');
+                $table->integer('project_id')->unsigned()->index();
+                $table->foreign('project_id')->references('id')->on('projects')->onDelete('cascade');
+                $table->integer('user_id')->unsigned()->index();
+                $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+                $table->timestamps();
+            });
+        }
+
+
 	}
